decoupled AnalogWave class from simulation
